Less code, more tests
