+++ giskard/llm/evaluators/plausibility.py
@@ -1,9 +1,4 @@
-import json
-
-from ...datasets.base import Dataset
-from ...models.base.model import BaseModel
-from .. import utils
-from .base import EVALUATE_MODEL_FUNCTIONS, EvaluationResult, LLMBasedEvaluator
+from .base import LLMBasedEvaluator
 
 PLAUSIBILITY_EVAL_PROMPT = """
 You are auditing an AI model. You are auding the following model:
@@ -30,49 +25,3 @@ If you are not sure, just answer 'I don’t know'.
 
 class PlausibilityEvaluator(LLMBasedEvaluator):
     _default_eval_prompt = PLAUSIBILITY_EVAL_PROMPT
-
-    def _make_evaluate_prompt(self, model: BaseModel, input_vars, model_output):
-        return self.eval_prompt.format(
-            model_name=model.meta.name,
-            model_description=model.meta.description,
-            input_vars=input_vars,
-            model_output=model_output,
-        )
-
-    def evaluate(self, model: BaseModel, dataset: Dataset):
-        model_outputs = model.predict(dataset).prediction
-
-        succeded = []
-        failed = []
-        errored = []
-
-        for input_vars, model_output in zip(
-            dataset.df.loc[:, model.meta.feature_names].to_dict("records"), model_outputs
-        ):
-            prompt = self._make_evaluate_prompt(model, input_vars, model_output)
-            out = utils.llm_fn_call(
-                [{"role": "system", "content": prompt}],
-                functions=EVALUATE_MODEL_FUNCTIONS,
-                temperature=self.llm_temperature,
-                model=self.llm_model,
-            )
-
-            try:
-                args = json.loads(out.function_call.arguments)
-                sample = {
-                    "input_vars": input_vars,
-                    "model_output": model_output,
-                    "reason": args.get("reason"),
-                }
-                if args["passed_test"]:
-                    succeded.append(sample)
-                else:
-                    failed.append(sample)
-            except (AttributeError, json.JSONDecodeError, KeyError):
-                errored.append({"input_vars": input_vars, "model_output": model_output})
-
-        return EvaluationResult(
-            failure_examples=failed,
-            success_examples=succeded,
-            errors=errored,
-        )
