+++ giskard/llm/evaluators/requirements.py
@@ -1,11 +1,7 @@
 from typing import Sequence
 
-from ...datasets.base import Dataset
 from ...models.base.model import BaseModel
-from ..client import LLMClient
-from ..client import llm_client as default_llm_client
-from ..errors import LLMGenerationError
-from .base import EVALUATE_MODEL_FUNCTIONS, EvaluationResult
+from .base import LLMBasedEvaluator
 
 EVALUATE_PROMPT = """Your role is to test AI models. Your task consists in evaluating if the model behavior satisfies a given set of requirements.
 
@@ -34,56 +30,18 @@ Otherwise, the model passes the test. This is the safest choice. If you are in d
 """
 
 
-class RequirementEvaluator:
-    def __init__(self, requirements: Sequence[str], llm_client: LLMClient = None):
-        self.requirements = requirements
-        self.llm_client = llm_client if llm_client is not None else default_llm_client
+class RequirementEvaluator(LLMBasedEvaluator):
+    _default_eval_prompt = EVALUATE_PROMPT
 
-    def _make_evaluate_prompt(self, model, input_vars, model_output):
-        return EVALUATE_PROMPT.format(
+    def __init__(self, requirements: Sequence[str], *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.requirements = requirements
+
+    def _make_evaluate_prompt(self, model: BaseModel, input_vars, model_output):
+        return self.eval_prompt.format(
             model_name=model.meta.name,
             model_description=model.meta.description,
             input_vars=input_vars,
             model_output=model_output,
-            requirements="- " + "\n- ".join(self.requirements),
-        )
-
-    def _make_evaluate_functions(self):
-        return EVALUATE_MODEL_FUNCTIONS
-
-    def evaluate(self, model: BaseModel, dataset: Dataset):
-        model_outputs = model.predict(dataset).prediction
-
-        succeded = []
-        failed = []
-        errored = []
-        for input_vars, model_output in zip(
-            dataset.df.loc[:, model.meta.feature_names].to_dict("records"), model_outputs
-        ):
-            sample = {"input_vars": input_vars, "model_output": model_output}
-            prompt = self._make_evaluate_prompt(model, input_vars, model_output)
-            funcs = self._make_evaluate_functions()
-            try:
-                out = self.llm_client.complete(
-                    [{"role": "system", "content": prompt}],
-                    functions=funcs,
-                    function_call={"name": "evaluate_model"},
-                    temperature=0.1,
-                )
-                if out.function_call is None or "passed_test" not in out.function_call.args:
-                    raise LLMGenerationError("Invalid function call arguments received")
-            except LLMGenerationError as err:
-                errored.append({"message": str(err), "sample": sample})
-                continue
-
-            args = out.function_call.args
-            if args["passed_test"]:
-                succeded.append({"input_vars": input_vars, "model_output": model_output, "reason": args.get("reason")})
-            else:
-                failed.append({"input_vars": input_vars, "model_output": model_output, "reason": args.get("reason")})
-
-        return EvaluationResult(
-            failure_examples=failed,
-            success_examples=succeded,
-            errors=errored,
+            requirements="\n".join([f"- {r}" for r in self.requirements]),
         )
